+++ src/main/webapp/views/mail/message.jsp
@@ -41,7 +41,7 @@
 			<option value="${box.id}"><jstl:out value="${box.name}"/></option>
 		</jstl:forEach>
 	</select>
-	<input type="submit" name = "move" value="<spring:message code="mail.move"/>"/>
+	<input type="submit" name = "move" value="<spring:message code="mail.move"/>" onclick="return confirm('<spring:message code="mail.move" />')"/>
 </form>
 
 <spring:url var="replyMessage" value="/message/authenticated/new.do?messageId={messageId}">
@@ -53,7 +53,7 @@
 
 <ul>
 	<li><a href="${replyMessage}"><spring:message code="mail.message.reply"/></a></li>
-	<li><a href="${deleteMessage}"><spring:message code="mail.message.delete"/></a></li>
+	<li><a href="${deleteMessage}" onclick="return confirm('<spring:message code="mail.delete" />')"><spring:message code="mail.message.delete"/></a></li>
 </ul>
 
 <p><jstl:out value="${message.body}"/></p>
